Routes can be added with a specific app
Overloads on routes take an AppDelegate or Action<IAppBuilder> parameter

App is passed forward as _app field. The OwinModule, when configured to
handle all request, will pass null for _app field.

Null _app field will default to the AppSingleton.Instance. That is the
app which is configured via the class Startup method.